refactor: repository에서 사용되는 데이터이므로 패키지 이동
