+++ android/app/src/main/java/app/priceguard/ui/additem/link/RegisterItemLinkViewModel.kt
@@ -3,9 +3,9 @@ package app.priceguard.ui.additem.link
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.priceguard.data.dto.verify.ProductVerifyRequest
-import app.priceguard.data.network.ProductErrorState
-import app.priceguard.data.network.ProductRepositoryResult
+import app.priceguard.data.repository.ProductErrorState
 import app.priceguard.data.repository.ProductRepository
+import app.priceguard.data.repository.ProductRepositoryResult
 import app.priceguard.ui.data.ProductVerifyResult
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
