+++ android/app/src/main/java/app/priceguard/ui/detail/ProductDetailViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.viewModelScope
 import app.priceguard.data.GraphDataConverter
 import app.priceguard.data.graph.ProductChartData
 import app.priceguard.data.graph.ProductChartDataset
-import app.priceguard.data.network.ProductErrorState
-import app.priceguard.data.network.ProductRepositoryResult
+import app.priceguard.data.repository.ProductErrorState
 import app.priceguard.data.repository.ProductRepository
+import app.priceguard.data.repository.ProductRepositoryResult
 import app.priceguard.materialchart.data.GraphMode
 import dagger.hilt.android.lifecycle.HiltViewModel
 import java.text.NumberFormat
