+++ android/app/src/main/java/app/priceguard/ui/home/list/ProductListViewModel.kt
@@ -3,9 +3,9 @@ package app.priceguard.ui.home.list
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import app.priceguard.data.GraphDataConverter
-import app.priceguard.data.network.ProductErrorState
-import app.priceguard.data.network.ProductRepositoryResult
+import app.priceguard.data.repository.ProductErrorState
 import app.priceguard.data.repository.ProductRepository
+import app.priceguard.data.repository.ProductRepositoryResult
 import app.priceguard.materialchart.data.GraphMode
 import app.priceguard.ui.home.ProductSummary.UserProductSummary
 import dagger.hilt.android.lifecycle.HiltViewModel
