+++ android/app/src/main/java/app/priceguard/ui/signup/SignupViewModel.kt
@@ -3,9 +3,9 @@ package app.priceguard.ui.signup
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.priceguard.data.network.AuthErrorState
-import app.priceguard.data.network.AuthRepositoryResult
+import app.priceguard.data.repository.AuthErrorState
 import app.priceguard.data.repository.AuthRepository
+import app.priceguard.data.repository.AuthRepositoryResult
 import app.priceguard.data.repository.TokenRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
